fixes folowing review
