+++ ocs_ci/helpers/helpers.py
@@ -1309,7 +1309,7 @@ def get_start_creation_time(interface, pvc_name):
     return datetime.datetime.strptime(start, format)
 
 
-def get_end_creation_time(interface, pvc_name, last_end_time=False):
+def get_end_creation_time(interface, pvc_name):
     """
     Get the ending creation time of a PVC based on provisioner logs
 
@@ -1332,23 +1332,18 @@ def get_end_creation_time(interface, pvc_name, last_end_time=False):
     logs = logs.split("\n")
     # Extract the starting time for the PVC provisioning
     end = [i for i in logs if re.search(f"provision.*{pvc_name}.*succeeded", i)]
-    # End provisioning string may appear in logs several times. If last_end_time is False, first such string is taken
-    # and if last_end_time is True -- the last such string is taken
-    if not last_end_time:
-        end = end[0].split(" ")[1]
-    else:
-        end = end[len(end) - 1].split(" ")[1]
+    # End provisioning string may appear in logs several times, take here the latest one
+    end = end[-1].split(" ")[1]
     return datetime.datetime.strptime(end, format)
 
 
-def measure_pvc_creation_time(interface, pvc_name, last_end_time=False):
+def measure_pvc_creation_time(interface, pvc_name):
     """
     Measure PVC creation time based on logs
 
     Args:
         interface (str): The interface backed the PVC
         pvc_name (str): Name of the PVC for creation time measurement
-        last_end_time (bool) if creation end time appears several times, take the last (latest) one
 
     Returns:
         float: Creation time for the PVC
@@ -1356,7 +1351,7 @@ def measure_pvc_creation_time(interface, pvc_name, last_end_time=False):
     """
     start = get_start_creation_time(interface=interface, pvc_name=pvc_name)
     end = get_end_creation_time(
-        interface=interface, pvc_name=pvc_name, last_end_time=last_end_time
+        interface=interface, pvc_name=pvc_name
     )
     total = end - start
     return total.total_seconds()
